Deploy hook
I hosted the splash page somewhere, and configured the host to always
update to the latest development version.

This commit only announces the URL where this development version can be
played. If you also want to configure your host to auto-update to the
latest version of a repository, create a cgi script which pulls from
GitHub whenever the page is visited. Then, go to your GitHub project's
admin panel and add a WebHook with that page as the URL.