Cake refactoring!
Renamed Level_{Data,Model,View} to World, Room, and... Cake!

The new naming conventions are shorter, the class responsibilities are
more well-defined, and the public interfaces between the classes are
tighter. All this, plus cake! What else could you ask for?

Seriously, I just couldn't think of a proper name for the class which
creates and updates the Layers which represent the current Room.