Okay, Cake was a silly name
The "first_container, second_container" code really wanted to be
refactored, so I put it in Cake, which was already managing the layers
of those containers.

I then added methods to manipulate those containers, and it just looked
silly to be calling methods on a cake. I thought long and hard about
what a proper name for this should be, and I found one: "Scene".

So, to recap, we have a World of Rooms of Tiles, and that room is
displayed in a Scene of Layers of Sprites. Sounds good?

I really like one-word class names.