Fixed: Cmd+R was reloading the level, not the page
It was annoying that keys such as Tab would apply to the browser instead
of the game, but it's even more annoying for keys like Cmd+R to apply to
the game instead of the browser.

We are now more careful in which keys are passed on to the browser: we
used to block everything, now we let letters through.