Detail: ground tiles stand out too much
In the first two levels, most of the screen is covered with wall tiles.
This makes the ground tiles below them stand out, because they have a
different color.

The ground tiles are not important. That's not what the player should be
focussing on. Therefore, I now detect this special case and remove the
ground tiles. You won't even notice they are gone.