Tracking Entities as they move around the room
When the room is represented solely as a grid of tiles, there is no
choice but to display it as a list of sprites.

Now, blocks and players are represented *both* as tiles and entities.
The fact that they are still represented as tiles will allow them to be
considered as obstacles for the other entities trying to occupy their
place.

And thanks to the fact that they are also represented by an entity --
some object which remains associated with a particular block or player
-- we finally have a persistent object to track as it moves around.