Colored blocks
There are now four types of blocks: regular, green, blue, and orange.
There are also four button types: regular, green, blue, and orange.

The idea is that regular (red) buttons accept any block, while colored
buttons only accept blocks of their color. This is not implemented yet,
but this will allow more complicated puzzles to be crafted.

Since there are now many types of block, it no longer useful to compare
a tile type with Tile.block; instead, you should check tile.entity,
which is a close approximation.

'g' is the symbol for green blocks, 'G' is the symbol for the green
button, and similarly for other colors. In particular, regular blocks
are now 'r', and the regular button is now 'R'.