Big refactoring: eschew callbacks
Instead of following the jQuery convention of registering and triggering
callbacks, let's follow the advice of Rich Hickey and Evan Czaplicki and
avoid using callbacks.

In the new design, data objects don't accumulate callbacks, but they do
accumulate events, which can be queried from them. Thus, control flow
never passes from data objects to view objects. Instead, the top level
(octocarina.js) controls both data and view objects, and passes the
events along when appropriate.

But is this new design better? Yes, I think so. Here is an example where
we can easily end up with a bug if the order of the callbacks isn't
carefully orchestrated.

I plan to refactor the code for switches into a different file, which
would need to listen to move events. The problem happens when we move a
block off a switch: two events are generated, and the order in which
they are generated is important. One of the events moves the block off
the switch, and other one moves the player on the switch in its place.

We need to make sure to trigger those two events in that particular
order: first the block, then the player. In this order, the switch first
deactivates, then quickly reactivates with the next event. With the
opposite order, the switch would stay deactivated after the two events.

If instead we implement this using event queues, it should be possible
to examine the entire list of events, detect the conflict, and do
something about it.

But really, the most important reason for making this change is that
code without callbacks is supposed to be easier to read. Is it?

You tell me.