More uniform class initialization
The multibuttons, forkedblock, and theatre variables are all
reinstantiated once the new Multiroom is created, but they were all
instantiated in a very different way. With this refactoring, they are
all instantiated via a constructor taking the current room as a
parameter (and a few more in the case of the Theatre).