Action queues instead of callbacks
We have already removed callbacks from the game objects like Room and
Multiroom, but there are still more to be removed. During the animations
occuring when we change level, for example, we want to suspend the
player movement, so we used callbacks to update a movement flag. This
commit gets rid of those callbacks by updating the flag in a different
way.

We introduce action queues, objects which know whether an animation is
under way. Graphical objects are no longer responsible for calling
callbacks, but they are now reponsible for maintaining the state of
their action queue whenever they start an animation.

Action queues know how to order actions and how to interact with other
action queues, allowing complex action sequences to be specified in a
declarative manner instead of relying on a complex mesh of callbacks
producing the correct order.