Lower sprite for the walls, closes #20
The reason the walls were so high to begin with is that their height was
the double of the block height, making them easy to stack and align.
Since we never stack blocks, this is not a feature we need to keep.

As a bonus, the door sprite is now the same height as the walls.