Lover explores ahead
I needed a reasonably-hard level without a button, so that the lover
could leave through the door but the player couldn't follow immediately.
This is what I came up with.

At first I thought it was impossible to solve, but after the Sokoban
level checker said it wasn't, I tried harder and I found a solution.

I am glad the level checker has finally been useful for something!