Easier levels first.
I needed a non-fork level in which the lover could screw up, but the
Players kept trying to use the fork to solve the "Mistake" level.

Hopefully, "Big pyramid" will make it clearer that the fork is not
needed, because the lover is not readily accessible.

I then placed the first fork level immediately afterwards, because
players expect fork-requiring levels early in the game.