Pushed everything on gelisam/game-off-2012
I read that only direct forks of github/game-off-2012 could qualify for
the contest. Since this team repo is a fork of my gelisam/game-off-2012
repo, this repo doesn't qualify!

Thus, I am pushing everything back to the parent repo, and I'll share
the prize with the rest of the team if that repo wins.

Closes #120.